Move new test case into EntitiesFinderSpec
